[closes #3087] Divide HTTP tests into separate modules
