+++ integration-tests/http/common/src/main/java/org/apache/camel/quarkus/component/http/common/HttpService.java
@@ -1,0 +1,60 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.camel.quarkus.component.http.common;
+
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.ws.rs.Consumes;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.POST;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
+import org.jboss.resteasy.annotations.GZIP;
+
+@Path("/service/common")
+@ApplicationScoped
+public class HttpService {
+    @Path("/get")
+    @GET
+    @Produces(MediaType.TEXT_PLAIN)
+    public String get() {
+        return "get";
+    }
+
+    @Path("/toUpper")
+    @POST
+    @Consumes(MediaType.TEXT_PLAIN)
+    @Produces(MediaType.TEXT_PLAIN)
+    public String toUpper(String message) {
+        return message.toUpperCase();
+    }
+
+    @Path("/https")
+    @GET
+    @Produces(MediaType.TEXT_PLAIN)
+    public String httpsGet() {
+        return "HTTPS GET";
+    }
+
+    @Path("/compress")
+    @GET
+    @GZIP
+    @Produces(MediaType.TEXT_PLAIN)
+    public String compress() {
+        return "Compressed response";
+    }
+}
