+++ integration-tests/http/common/src/main/resources/jsse/truststore.jks
