+++ integration-tests/http/netty-http/src/test/java/org/apache/camel/quarkus/component/http/netty/it/NettyHttpTest.java
@@ -1,0 +1,74 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.camel.quarkus.component.http.netty.it;
+
+import io.quarkus.test.common.QuarkusTestResource;
+import io.quarkus.test.junit.QuarkusTest;
+import io.restassured.RestAssured;
+import org.apache.camel.quarkus.component.http.common.AbstractHttpTest;
+import org.apache.camel.quarkus.component.http.common.HttpTestResource;
+import org.eclipse.microprofile.config.ConfigProvider;
+import org.junit.jupiter.api.Test;
+
+import static org.hamcrest.Matchers.is;
+
+@QuarkusTest
+@QuarkusTestResource(HttpTestResource.class)
+public class NettyHttpTest extends AbstractHttpTest {
+    @Override
+    public String component() {
+        return "netty-http";
+    }
+
+    @Test
+    public void basicNettyHttpServer() {
+        RestAssured
+                .given()
+                .port(ConfigProvider.getConfig().getValue("camel.netty-http.test-port", Integer.class))
+                .when()
+                .get("/test/server/hello")
+                .then()
+                .statusCode(200)
+                .body(is("Netty Hello World"));
+    }
+
+    @Test
+    public void transferException() {
+        RestAssured
+                .given()
+                .queryParam("test-port", getPort("camel.netty-http.test-port"))
+                .when()
+                .get("/test/client/{component}/serialized/exception", component())
+                .then()
+                .statusCode(200)
+                .body(is("java.lang.IllegalStateException"));
+    }
+
+    @Override
+    @Test
+    public void compression() {
+        final int port = getPort("camel.netty-http.compression-test-port");
+        RestAssured
+                .given()
+                .queryParam("test-port", port)
+                .when()
+                .get("/test/client/{component}/compression", component())
+                .then()
+                .statusCode(200)
+                .body(is("Netty Hello World Compressed"));
+    }
+}
